requirements.txt
PyYAML
opencv-python
easyocr
pytesseract
torch  # Required by easyocr